fix an issue with account public keys
